refactor(economy): extract the economy engine into services
