+++ src/discordbot/cogs/_games/lobby.py
@@ -14,8 +14,8 @@ from nextcord.ui import Item, View, Button
 from discordbot.typings.economy import JackpotSettlementRequest, JackpotSettlementBatchResult
 from discordbot.utils.discord_embeds import embed_spacer_payload
 from discordbot.utils.message_cleanup import schedule_public_message_delete
-from discordbot.cogs._economy.database import apply_jackpot_settlement_batch
 from discordbot.cogs._games.interactions import disable_view_components
+from discordbot.services.economy.database import apply_jackpot_settlement_batch
 from discordbot.utils.owned_message_views import send_ephemeral_notice
 
 if TYPE_CHECKING:
