+++ tests/test_economy.py
@@ -20,7 +20,8 @@ from discordbot.typings.games import (
 from discordbot.utils.timezone import TAIWAN_TIMEZONE
 from discordbot.typings.economy import TRANSFER_TAX_BPS
 from discordbot.cogs._games.blackjack import Card, BlackjackRound, BlackjackHandState
-from discordbot.cogs._economy.database import (
+from discordbot.cogs._games.settlement import settle_wager, settle_blackjack_player
+from discordbot.services.economy.database import (
     VIP_PURCHASE_COST,
     CHECKIN_STREAK_CYCLE,
     BASE_CHECKIN_REWARD_AMOUNT,
@@ -66,7 +67,6 @@ from discordbot.cogs._economy.database import (
     _apply_daily_casino_delta_in_session,
     invalidate_economy_leaderboard_cache,
 )
-from discordbot.cogs._games.settlement import settle_wager, settle_blackjack_player
 from discordbot.cogs._games.blackjack_views import BlackjackView, build_final_embeds
 
 from tests.helpers.casting import as_message, as_interaction
@@ -473,8 +473,8 @@ async def test_ensure_schema_bootstraps_current_databases(
     """A clean startup creates only the current economy tables."""
     db_path = tmp_path / "current-economy.db"
     engine = create_async_engine(url=f"sqlite+aiosqlite:///{db_path}")
-    monkeypatch.setattr("discordbot.cogs._economy.database._engine", engine)
-    monkeypatch.setattr("discordbot.cogs._economy.database._schema_ready_for", None)
+    monkeypatch.setattr("discordbot.services.economy.database._engine", engine)
+    monkeypatch.setattr("discordbot.services.economy.database._schema_ready_for", None)
 
     await _ensure_schema()
 
@@ -528,8 +528,8 @@ async def test_ensure_schema_serializes_concurrent_first_use(
     """Concurrent first-use schema bootstrap does not race SQLite CREATE TABLE."""
     db_path = tmp_path / "concurrent-economy.db"
     engine = create_async_engine(url=f"sqlite+aiosqlite:///{db_path}")
-    monkeypatch.setattr("discordbot.cogs._economy.database._engine", engine)
-    monkeypatch.setattr("discordbot.cogs._economy.database._schema_ready_for", None)
+    monkeypatch.setattr("discordbot.services.economy.database._engine", engine)
+    monkeypatch.setattr("discordbot.services.economy.database._schema_ready_for", None)
 
     await asyncio.gather(*(_ensure_schema() for _ in range(20)))
 
@@ -2339,7 +2339,7 @@ async def test_apply_jackpot_settlement_batch_rolls_back_on_failure(
         return await original_apply(**kwargs)
 
     monkeypatch.setattr(
-        "discordbot.cogs._economy.database._apply_jackpot_delta_in_session",
+        "discordbot.services.economy.database._apply_jackpot_delta_in_session",
         flaky_apply_jackpot_delta_in_session,
     )
 
@@ -2405,15 +2405,15 @@ async def test_ensure_schema_seeds_dragon_gate_jackpot_once(
     """_ensure_schema seeds the dragon_gate pool exactly once across calls."""
     db_path = tmp_path / "seed-economy.db"
     engine = create_async_engine(url=f"sqlite+aiosqlite:///{db_path}")
-    monkeypatch.setattr("discordbot.cogs._economy.database._engine", engine)
-    monkeypatch.setattr("discordbot.cogs._economy.database._schema_ready_for", None)
+    monkeypatch.setattr("discordbot.services.economy.database._engine", engine)
+    monkeypatch.setattr("discordbot.services.economy.database._schema_ready_for", None)
 
     await _ensure_schema()
     first_balance = await get_jackpot_pool(game_id="dragon_gate")
     assert first_balance == 1_000
 
     # Calling again is idempotent: the seed must not pile on top of itself.
-    monkeypatch.setattr("discordbot.cogs._economy.database._schema_ready_for", None)
+    monkeypatch.setattr("discordbot.services.economy.database._schema_ready_for", None)
     await _ensure_schema()
     assert await get_jackpot_pool(game_id="dragon_gate") == 1_000
